remote.c (remote_crc_query): Fix argument type mismatch with its declaration.
From: Jie Zhang <jie.zhang@analog.com>

Builds on 64bit systems will fail like so:
remote.c:1458: error: conflicting types for ‘remote_crc_query’
remote.c:145: note: previous declaration of ‘remote_crc_query’ was here

+++ remote.c
@@ -1456,7 +1456,7 @@ static int remote_offsets_query(uint64_t *text,
 
 /* Target method */
 static int remote_crc_query(uint64_t addr,
-                            unsigned int len,
+                            size_t len,
                             uint32_t *val)
 {
     char in_buf[RP_PARAM_INOUTBUF_SIZE];
